add initial event listener component
The event listener is very basic at this stage. It has only the on death event that you can get a callback from. There are a couple of design decisions that might seem strange but have a reason behind them. First of all when you look at the component you see an eid in the definition, that is because the listener is not attached to another entity as a component, but rather uses its own entity, as seen in the helper method `addEventListenerEntity`. This is because I want to be able to register multiple different listener to a single entity. For example you might want to spawn items on death & show an explosion, if you could only register one listener you would have to always combine these in some way and you would not be able to abstract them away that easily. The downside of this approach is that you have to keep track of the listener entities and delete them once the entities they are referencing are gone.